requirements.txt
langchain==0.0.325
python-dotenv==1.0.1
openai==0.28.1 